requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pyyaml==6.0.1
flask==3.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==6.1.0
black==23.11.0
